Use Windows newlines in the wifi.yaml file
We encourage teams to open this file, yet as most use Windows they
will open it in Notepage and find that it looks odd. In turn they
are likely to input the wrong WiFi passphrase as a result. This
gives them a bad experience and takes up our time in diagnosing
the issue and helping them.

If we output the file with Windows newlines then all the above
issues go away. This should be safe for non-Windows users too,
since *nix and mac tools are generally tolerant to Windows newlines.

Patch set 2: use the 'io' module to get an 'open' which supports
             the 'newline' argument.

Change-Id: I41d48e2be9c7bd9d541d73d05d277318ee8c58dc